examples: libmetal: irq_shmem_demo: move channel state to common header
Move the shared irq_shmem_demo channel definition into a demo-level
header and keep platform-specific state behind a machine_ctx pointer.

This removes the duplicated channel_s definitions from the host and
remote OS headers and makes the separation between common transport
state and platform-private state clearer.

Also export the demo common include directory through CMake so the
shared header can be included directly as irq_shmem_demo.h instead of
using deep relative include paths.
