+++ examples/libmetal/demos/irq_shmem_demo/CMakeLists.txt
@@ -1,3 +1,4 @@
 # Copyright (C) 2025 Advanced Micro Devices, Inc.  All rights reserved.
 # SPDX-License-Identifier: BSD-3-Clause
+collect (PROJECT_INC_DIRS "${CMAKE_CURRENT_SOURCE_DIR}/common")
 add_subdirectory(${ROLE})
