+++ examples/libmetal/machine/remote/amd_rpu/system/generic/amp_demo_os.h
@@ -14,18 +14,20 @@
 #include <metal/cpu.h>
 #include <metal/sys.h>
 
+#include "irq_shmem_demo.h"
 #include "xil_printf.h"
 
-struct channel_s {
-	struct metal_io_region *host_to_remote_desc_io; /* host to remote descriptors */
-	struct metal_io_region *remote_to_host_desc_io; /* remote to host descriptors */
-	struct metal_io_region *ipi_io; /* IPI metal i/o region */
-	struct metal_io_region *shm_io; /* Shared memory metal i/o region */
-	struct metal_io_region *ttc_io; /* TTC metal i/o region */
-	uint32_t ipi_mask;              /* RPU IPI mask */
-	int irq_vector_id;              /* IRQ number. */
+struct channel_machine_ctx_s {
 	atomic_flag irq_pending;        /* Lightweight wait primitive. */
 };
+
+static inline struct channel_machine_ctx_s *channel_machine_ctx(struct channel_s *ch)
+{
+	metal_assert(ch);
+	metal_assert(ch->machine_ctx);
+
+	return (struct channel_machine_ctx_s *)ch->machine_ctx;
+}
 
 /**
  * @brief amp_os_init() - initialize bare-metal rendezvous primitives
@@ -38,14 +40,14 @@ struct channel_s {
  */
 static inline int amp_os_init(struct channel_s *ch, void *arg)
 {
+	struct channel_machine_ctx_s *machine = channel_machine_ctx(ch);
+
 	(void)arg;
 
-	metal_assert(ch);
-
-	ch->irq_pending = (atomic_flag)ATOMIC_FLAG_INIT;
+	machine->irq_pending = (atomic_flag)ATOMIC_FLAG_INIT;
 
 	/* Start in the "waiting" state until the first interrupt arrives. */
-	(void)atomic_flag_test_and_set(&ch->irq_pending);
+	(void)atomic_flag_test_and_set(&machine->irq_pending);
 
 	return 0;
 }
@@ -60,9 +62,9 @@ static inline int amp_os_init(struct channel_s *ch, void *arg)
  */
 static inline void system_suspend(struct channel_s *ch)
 {
-	metal_assert(ch);
+	struct channel_machine_ctx_s *machine = channel_machine_ctx(ch);
 
-	while (atomic_flag_test_and_set(&ch->irq_pending)) {
+	while (atomic_flag_test_and_set(&machine->irq_pending)) {
 		metal_asm volatile("wfi");
 	}
 }
@@ -74,8 +76,9 @@ static inline void system_suspend(struct channel_s *ch)
  */
 static inline void system_resume(struct channel_s *ch)
 {
-	metal_assert(ch);
-	atomic_flag_clear(&ch->irq_pending);
+	struct channel_machine_ctx_s *machine = channel_machine_ctx(ch);
+
+	atomic_flag_clear(&machine->irq_pending);
 }
 
 #endif /* __AMP_DEMO_OS_H__ */
